Split knit land.rs into a land/ module
Break the 2,454-line src/commands/land.rs into focused submodules:
plan (build the default plan), validate (validate + preflight PRs),
execute (schedule and run the plan), update (knit land update), and
display (render plans/run status). land/mod.rs keeps the public command
entrypoints, shared plan/run types and constants, and cross-cutting helpers.
No behavior change; tests and clippy green.


Knit-Group: kg_20260530_fcbf59
Knit-Bundle: split-oversized-source-files
